Generalise filesystem mmap-safety check beyond 9p
The previous startup check refused to start QuestDB on 9p mounts unless
commit.mode=sync. The same hazard exists on VirtIO-FS (Docker Desktop on
macOS), generic FUSE drivers, and SMB shares: mmap'd writes are buffered
in a client page cache and may silently never reach the underlying file
without an explicit msync. Inverting the check to a per-OS mmap-safe
whitelist makes the policy correct for all of these without needing a
new entry per hostile filesystem.

Native getFileSystemStatus now packs three flag bits into the high half
of the jlong return value: FLAG_FS_SUPPORTED (recognised by name),
FLAG_FS_MMAP_SAFE (mmap+writeback durable without msync), FLAG_FS_HARD_FAIL
(refuse to start regardless of commit.mode).

Per-OS whitelist:
  Linux:   ext4, xfs, zfs, overlayfs, tmpfs are mmap-safe.
           v9fs, fuse, cifs, smb, smb2 are recognised but not mmap-safe.
           nfs hard-fails.
           v9fs has been demoted from the supported list (was a bug).
  macOS:   apfs is mmap-safe (matched by f_fstypename and f_type).
           virtiofs, smbfs, fuse, macfuse are recognised but not mmap-safe.
           nfs hard-fails.
           Removes the empirical 0x6a656a63 magic check.
  Windows: GetDriveType + GetVolumeInformation distinguish local from
           remote. Local NTFS/ReFS are mmap-safe; local FAT family and
           SMB-redirected NTFS/ReFS require sync; other DRIVE_REMOTE
           hard-fails. Replaces the misleading FILE_SUPPORTS_TRANSACTIONS
           heuristic which marked SMB-served NTFS as supported.
  FreeBSD: ufs (0x35), zfs, tmpfs are mmap-safe.

Bootstrap.checkV9fsCommitMode is replaced by checkMmapSafeOrSync, which
returns one of three outcomes per data-path root: hard-fail (NFS,
unrecognised remote share), mmap-safe (proceed), or require sync.

Operators can override the sync requirement by prefixing the value with
'!' (commit.mode=!nosync). The override emits a loud advisoryW warning
and does NOT bypass hard-fail filesystems.

The startup log now distinguishes four states (INCOMPATIBLE, SUPPORTED,
SUPPORTED REQUIRES commit.mode=sync, UNSUPPORTED) instead of the prior
binary SUPPORTED / UNSUPPORTED.

VIRTIO_FS_DETECTED used by TableUtils.lock for the Docker macOS touch
workaround now reads the path buffer name "virtiofs" written by the
native side, replacing the empirical 0x6a656a63 magic which is no
longer set.

Includes a new windows-refs-bootstrap CI matrix entry that creates a
ReFS-formatted VHDX via diskpart, mounts it as R:, and runs Bootstrap
plus core storage tests against it. The diskpart sequence uses /s with
an ascii-encoded script file rather than stdin piping.

+++ core/src/main/java/io/questdb/std/FilesFacade.java
@@ -78,7 +78,7 @@ public interface FilesFacade {
 
     long getFileLimit();
 
-    int getFileSystemStatus(LPSZ lpszName);
+    long getFileSystemStatus(LPSZ lpszName);
 
     long getLastModified(LPSZ path);
